[ES6] Use temp allocation index per scope
We started to receive lint warning about "var redeclaration". This was due temp allocated vars reused temp index, that
produced several `var $__0`, `var $__0` statements.

Technically it wasn't an issue for the real code, mostly for the JSHint linter, however, instead of tweaking the linter, I decided to go in the
correct way, and use temp vars index per scope.

The `tempVarIndex` property in the `localScope` of the state stores the current temp index. This can be reused in other transforms that will require
usage of temp vars.