[JS] Trailing comma in Array and Object Initializers
People want to write trailing commas in Array and Object initializer to be consistent with PHP in this case. In prod, we already have a
minifier that strips the trailing comma, but not in dev, that causes problems of testing in IE.

This is the follow up of stripping commas, and Array and Object literals, but only those which have the trailing comma.

Alternatively, if we could configure the comma removal for dev, that would probably better that this transform. Though, if we don't have an alternative here, the transform is the way to go.