Support ES7 rest properties in destructuring
This adds support for ES7 rest properties in the ES6 destructuring. This is
kind of lame since the ES6 destructuring takes on a dependency on the ES7 one
but it simplifies things since the ES6 destructuring isn't really composable.

Currently the helper function is inlined on every use. I don't expect this to
be used very much yet so this is probably fine until we have a runtime or
recast.