Silently ignore null/undefined source arguments to Object.assign.
This week's TC39 committee meeting decided in favor of not throwing when `null` or `undefined` arguments are passed to `Object.assign`. Instead, they are silently ignored, as if they were empty objects. Note, however, that it is still an error for the first argument (the target object) to be `null`.