Fix issue in call-spread visitors where the AST was being mutated
If you simply parse one file and throw away the AST, this isn't really a
problem. But there are use cases where the AST is cached +
re-used...meaning that if you mutate the AST, you effectively poison the
cache for later uses.

This just fixes the transform to NOT mutate the AST.