Add the new API and CLI
This mostly behaves the same way react-tools did. One big difference is
that the transform function exported does not have a simple string
return value. This is more like transformWithDetails.

This is inline with how babel does it, and I like it. We already have
the information, we might as well expose it and leave it to the consumer
to get what it wants off.