Fix call spread transform with temp vars
The root of the problem is that we never caught up to the beginning of
the function body. So our buffer never included the opening brace when
we stopped visiting the fn params. This works out most of the time
because we eventually catchup. But when we insert temp vars we assume
the buffer is caught up to the function body. But we didn't do that so
temp vars were inserted at the wrong place.